WIP making client
少し構成を見直し、大した修正ではないもの
