--- src/components/GamesList.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import GameGridList from './GameGridList';
-import './Main.css';
-
-export default () => (
-  <div className="Container">
-    <GameGridList />
-  </div>
-);
